+++ src/components/diagrams/singleLineDiagram/single-line-diagram-pane.js
@@ -19,12 +19,12 @@ import {
     getVoltageLevelSingleLineDiagram,
     updateSwitchState,
 } from '../../../utils/rest-api';
-import SingleLineDiagram, { SvgType } from './single-line-diagram';
+import SingleLineDiagram from './single-line-diagram';
 import PropTypes from 'prop-types';
 import { Chip, Stack } from '@mui/material';
 import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import makeStyles from '@mui/styles/makeStyles';
-import { getNameOrId, useSingleLineDiagram, ViewState } from './utils';
+import { getNameOrId, SvgType, useSingleLineDiagram, ViewState } from './utils';
 import { isNodeBuilt } from '../../graph/util/model-functions';
 import { AutoSizer } from 'react-virtualized';
 import { SLD_DISPLAY_MODE } from '../../network/constants';
@@ -52,7 +52,7 @@ const useDisplayView = (network, studyUuid, currentNode) => {
                       centerName,
                       diagonalName,
                       componentLibrary,
-                      SLD_DISPLAY_MODE.DEFAULT
+                      SLD_DISPLAY_MODE.STATE_VARIABLE
                   )
                 : null,
         [
