+++ src/components/diagrams/singleLineDiagram/single-line-diagram.js
@@ -53,19 +53,11 @@ import { useIsAnyNodeBuilding } from '../../util/is-any-node-building-hook';
 import Alert from '@mui/material/Alert';
 import { isNodeReadOnly } from '../../graph/util/model-functions';
 import { SingleLineDiagramViewer } from '@powsybl/diagram-viewer';
-
-export const SubstationLayout = {
-    HORIZONTAL: 'horizontal',
-    VERTICAL: 'vertical',
-    SMART: 'smart',
-    SMARTHORIZONTALCOMPACTION: 'smartHorizontalCompaction',
-    SMARTVERTICALCOMPACTION: 'smartVerticalCompaction',
-};
-
-export const SvgType = {
-    VOLTAGE_LEVEL: 'voltage-level',
-    SUBSTATION: 'substation',
-};
+import {
+    computePaperAndSvgSizesIfReady,
+    setWidthAndHeight,
+    SvgType,
+} from './utils';
 
 const loadingWidth = 150;
 const maxWidthVoltageLevel = 800;
@@ -184,51 +176,6 @@ let initialWidth, initialHeight;
 const mapRightOffset = 120;
 const mapBottomOffset = 80;
 const borders = 2; // we use content-size: border-box so this needs to be included..
-// Compute the paper and svg sizes. Returns undefined if the preferred sizes are undefined.
-const computePaperAndSvgSizesIfReady = (
-    fullScreen,
-    svgType,
-    totalWidth,
-    totalHeight,
-    svgPreferredWidth,
-    svgPreferredHeight,
-    headerPreferredHeight
-) => {
-    if (
-        typeof svgPreferredWidth != 'undefined' &&
-        typeof headerPreferredHeight != 'undefined'
-    ) {
-        let paperWidth, paperHeight, svgWidth, svgHeight;
-        if (fullScreen) {
-            paperWidth = totalWidth;
-            paperHeight = totalHeight;
-            svgWidth = totalWidth - borders;
-            svgHeight = totalHeight - headerPreferredHeight - borders;
-        } else {
-            let maxWidth, maxHeight;
-            if (svgType === SvgType.VOLTAGE_LEVEL) {
-                maxWidth = maxWidthVoltageLevel;
-                maxHeight = maxHeightVoltageLevel;
-            } else {
-                maxWidth = maxWidthSubstation;
-                maxHeight = maxHeightSubstation;
-            }
-            svgWidth = Math.min(
-                svgPreferredWidth,
-                totalWidth - mapRightOffset,
-                maxWidth
-            );
-            svgHeight = Math.min(
-                svgPreferredHeight,
-                totalHeight - mapBottomOffset - headerPreferredHeight,
-                maxHeight
-            );
-            paperWidth = svgWidth + borders;
-            paperHeight = svgHeight + headerPreferredHeight + borders;
-        }
-        return { paperWidth, paperHeight, svgWidth, svgHeight };
-    }
-};
 
 const SingleLineDiagram = forwardRef((props, ref) => {
     const [svg, setSvg] = useState(noSvg);
@@ -334,9 +281,16 @@ const SingleLineDiagram = forwardRef((props, ref) => {
             totalHeight,
             svgPreferredWidth,
             svgPreferredHeight,
-            headerPreferredHeight
+            headerPreferredHeight,
+            borders,
+            maxWidthVoltageLevel,
+            maxHeightVoltageLevel,
+            maxWidthSubstation,
+            maxHeightSubstation,
+            mapRightOffset,
+            mapBottomOffset
         );
-        if (typeof sizes != 'undefined') {
+        if (sizes) {
             if (
                 !fullScreenSldId &&
                 sizes.svgWidth * numberToDisplay > totalWidth
@@ -654,28 +608,16 @@ const SingleLineDiagram = forwardRef((props, ref) => {
         );
     }
 
-    let sizeWidth,
-        sizeHeight = initialHeight;
-    if (svg.error) {
-        sizeWidth = errorWidth;
-    } else if (
-        typeof finalPaperWidth != 'undefined' &&
-        typeof finalPaperHeight != 'undefined'
-    ) {
-        sizeWidth = finalPaperWidth;
-        sizeHeight = finalPaperHeight;
-    } else if (initialWidth !== undefined || loadingState) {
-        sizeWidth = initialWidth;
-    } else {
-        sizeWidth = totalWidth; // happens during initialization if initial width value is undefined
-    }
-
-    if (sizeWidth !== undefined) {
-        initialWidth = sizeWidth; // setting initial width for the next SLD.
-    }
-    if (sizeHeight !== undefined) {
-        initialHeight = sizeHeight; // setting initial height for the next SLD.
-    }
+    let { sizeWidth, sizeHeight } = setWidthAndHeight(
+        svg,
+        finalPaperWidth,
+        finalPaperHeight,
+        loadingState,
+        totalWidth,
+        initialHeight,
+        initialWidth,
+        errorWidth
+    );
 
     const pinSld = useCallback(() => onTogglePin(sldId), [sldId, onTogglePin]);
